Revert "LPS-115709 Rename classes"
This reverts commit 9e37978b73cb26a5bb0728dfab681516ab741baf.
