+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/java/com/liferay/app/builder/workflow/web/internal/portlet/action/AddAppBuilderAppMVCResourceCommand.java
@@ -15,18 +15,20 @@
 package com.liferay.app.builder.workflow.web.internal.portlet.action;
 
 import com.liferay.app.builder.constants.AppBuilderPortletKeys;
-import com.liferay.app.builder.model.AppBuilderAppVersion;
+import com.liferay.app.builder.model.AppBuilderApp;
 import com.liferay.app.builder.rest.dto.v1_0.App;
 import com.liferay.app.builder.rest.resource.v1_0.AppResource;
-import com.liferay.app.builder.service.AppBuilderAppVersionLocalService;
 import com.liferay.app.builder.workflow.rest.dto.v1_0.AppWorkflow;
 import com.liferay.app.builder.workflow.rest.resource.v1_0.AppWorkflowResource;
+import com.liferay.dynamic.data.lists.model.DDLRecord;
+import com.liferay.portal.kernel.model.WorkflowDefinitionLink;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
+import com.liferay.portal.kernel.security.permission.ResourceActionsUtil;
+import com.liferay.portal.kernel.service.WorkflowDefinitionLinkLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 
-import java.util.Objects;
 import java.util.Optional;
 
 import javax.portlet.ResourceRequest;
@@ -40,11 +42,11 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	property = {
 		"javax.portlet.name=" + AppBuilderPortletKeys.APPS,
-		"mvc.command.name=/app_builder/update_workflow_app"
+		"mvc.command.name=/app_builder/add_workflow_app"
 	},
 	service = MVCResourceCommand.class
 )
-public class UpdateWorkflowAppMVCResourceCommand
+public class AddAppBuilderAppMVCResourceCommand
 	extends BaseAppBuilderAppMVCResourceCommand<App> {
 
 	@Override
@@ -60,13 +62,8 @@ public class UpdateWorkflowAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		App app = appResource.getApp(
-			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
-
-		String appVersion = app.getVersion();
-
-		app = appResource.putApp(
-			ParamUtil.getLong(resourceRequest, "appBuilderAppId"),
+		App app = appResource.postDataDefinitionApp(
+			ParamUtil.getLong(resourceRequest, "dataDefinitionId"),
 			App.toDTO(ParamUtil.getString(resourceRequest, "app")));
 
 		AppWorkflowResource appWorkflowResource = AppWorkflowResource.builder(
@@ -74,52 +71,28 @@ public class UpdateWorkflowAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		AppWorkflow appWorkflow = AppWorkflow.toDTO(
-			ParamUtil.getString(resourceRequest, "appWorkflow"));
+		appWorkflowResource.postAppWorkflow(
+			app.getId(),
+			AppWorkflow.toDTO(
+				ParamUtil.getString(resourceRequest, "appWorkflow")));
 
-		if (_isModifiedAppWorkflow(
-				appWorkflowResource.getAppWorkflow(
-					ParamUtil.getLong(resourceRequest, "appBuilderAppId")),
-				appWorkflow)) {
+		WorkflowDefinitionLink workflowDefinitionLink =
+			_workflowDefinitionLinkLocalService.getWorkflowDefinitionLink(
+				themeDisplay.getCompanyId(), 0,
+				ResourceActionsUtil.getCompositeModelName(
+					AppBuilderApp.class.getName(), DDLRecord.class.getName()),
+				app.getId(), 0);
 
-			if (Objects.equals(app.getVersion(), appVersion)) {
-				AppBuilderAppVersion appBuilderAppVersion =
-					_appBuilderAppVersionLocalService.addAppBuilderAppVersion(
-						app.getSiteId(), themeDisplay.getCompanyId(),
-						themeDisplay.getUserId(), app.getId(),
-						app.getDataRecordCollectionId(),
-						app.getDataDefinitionId(), app.getDataLayoutId());
-
-				app.setVersion(appBuilderAppVersion.getVersion());
-			}
-
-			appWorkflowResource.putAppWorkflow(app.getId(), appWorkflow);
-		}
+		app.setWorkflowDefinitionName(
+			workflowDefinitionLink.getWorkflowDefinitionName());
+		app.setWorkflowDefinitionVersion(
+			workflowDefinitionLink.getWorkflowDefinitionVersion());
 
 		return Optional.of(app);
 	}
 
-	private boolean _isModifiedAppWorkflow(
-		AppWorkflow appWorkflow, AppWorkflow newAppWorkflow) {
-
-		if (!Objects.deepEquals(
-				appWorkflow.getAppWorkflowStates(),
-				newAppWorkflow.getAppWorkflowStates())) {
-
-			return true;
-		}
-
-		if (!Objects.deepEquals(
-				appWorkflow.getAppWorkflowTasks(),
-				newAppWorkflow.getAppWorkflowTasks())) {
-
-			return true;
-		}
-
-		return false;
-	}
-
 	@Reference
-	private AppBuilderAppVersionLocalService _appBuilderAppVersionLocalService;
+	private WorkflowDefinitionLinkLocalService
+		_workflowDefinitionLinkLocalService;
 
 }
