+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/java/com/liferay/app/builder/workflow/web/internal/portlet/action/DeleteAppBuilderAppMVCResourceCommand.java
@@ -15,16 +15,9 @@
 package com.liferay.app.builder.workflow.web.internal.portlet.action;
 
 import com.liferay.app.builder.constants.AppBuilderPortletKeys;
-import com.liferay.app.builder.model.AppBuilderApp;
-import com.liferay.app.builder.rest.dto.v1_0.App;
 import com.liferay.app.builder.rest.resource.v1_0.AppResource;
-import com.liferay.app.builder.workflow.rest.dto.v1_0.AppWorkflow;
 import com.liferay.app.builder.workflow.rest.resource.v1_0.AppWorkflowResource;
-import com.liferay.dynamic.data.lists.model.DDLRecord;
-import com.liferay.portal.kernel.model.WorkflowDefinitionLink;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
-import com.liferay.portal.kernel.security.permission.ResourceActionsUtil;
-import com.liferay.portal.kernel.service.WorkflowDefinitionLinkLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
@@ -34,7 +27,6 @@ import java.util.Optional;
 import javax.portlet.ResourceRequest;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Rafael Praxedes
@@ -42,15 +34,15 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	property = {
 		"javax.portlet.name=" + AppBuilderPortletKeys.APPS,
-		"mvc.command.name=/app_builder/add_workflow_app"
+		"mvc.command.name=/app_builder/delete_workflow_app"
 	},
 	service = MVCResourceCommand.class
 )
-public class AddWorkflowAppMVCResourceCommand
-	extends BaseAppBuilderAppMVCResourceCommand<App> {
+public class DeleteAppBuilderAppMVCResourceCommand
+	extends BaseAppBuilderAppMVCResourceCommand<Void> {
 
 	@Override
-	protected Optional<App> doTransactionalCommand(
+	protected Optional<Void> doTransactionalCommand(
 			ResourceRequest resourceRequest)
 		throws Exception {
 
@@ -62,37 +54,18 @@ public class AddWorkflowAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		App app = appResource.postDataDefinitionApp(
-			ParamUtil.getLong(resourceRequest, "dataDefinitionId"),
-			App.toDTO(ParamUtil.getString(resourceRequest, "app")));
+		appResource.deleteApp(
+			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
 
 		AppWorkflowResource appWorkflowResource = AppWorkflowResource.builder(
 		).user(
 			themeDisplay.getUser()
 		).build();
 
-		appWorkflowResource.postAppWorkflow(
-			app.getId(),
-			AppWorkflow.toDTO(
-				ParamUtil.getString(resourceRequest, "appWorkflow")));
+		appWorkflowResource.deleteAppWorkflow(
+			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
 
-		WorkflowDefinitionLink workflowDefinitionLink =
-			_workflowDefinitionLinkLocalService.getWorkflowDefinitionLink(
-				themeDisplay.getCompanyId(), 0,
-				ResourceActionsUtil.getCompositeModelName(
-					AppBuilderApp.class.getName(), DDLRecord.class.getName()),
-				app.getId(), 0);
-
-		app.setWorkflowDefinitionName(
-			workflowDefinitionLink.getWorkflowDefinitionName());
-		app.setWorkflowDefinitionVersion(
-			workflowDefinitionLink.getWorkflowDefinitionVersion());
-
-		return Optional.of(app);
+		return Optional.empty();
 	}
-
-	@Reference
-	private WorkflowDefinitionLinkLocalService
-		_workflowDefinitionLinkLocalService;
 
 }
